Rewrote the DeviceConfig struct so that it uses pure Go types and is no longer directly passable to Miniaudio. This is to avoide subtle errors where Miniaudio updates, an internal struct field type changes, and suddenly your app is crashing for no good reason because the Malgo version of the Config struct is still using the old types, some field offsets are all wrong ETC. Before this, Malgo could crash on 32-bit Windows. Now CGO resolves struct field offsets and we just convert between C and Go.
DeviceConfig now uses anonymous structs just like Miniaudio. The advantage of this is that everything is in one place and looks nicer in docs, and we can remove a bunch of types that are only ever used in device config and clutter up the docs.

Added the DeviceConfig.toCRepr() method which initializes a C.ma_device_config and returns that, along with a pointerList containing any allocated memory. See the comments in DeviceConfig.toCRepr.

All fields with type uint32 are now type int, meaning you usually need less type conversions. Many Go functions just use int by default and converting between everything is annoying.

Some fields with type uint32 now have type bool, as appropriate.

Some fields with type *byte now have type string and have P removed from the beginning of their name as (one we're using Go convensions) and two (they're not pointers anymore), making it way easier to use them. Strings are allocated, stored and freed using a pointerList (see utils.go) and the user doesn't have to do anything.

PeriodSizeInMilliseconds is now PeriodSizeInDuration and has type time.Duration, making it easier to use.

We now use pointerList to properly clean up deviceID when a non-default device is requested and the Device is later freed.

The Device callbacks (currently DataProc and StopProc) must still be passed to InitDevice in a DeviceCallbacks struct, but there's now nothing stopping us from just making it a member of the config and making the InitDevice function easier to use while we're at it.

Breaking change because the types of nearly all if not all fields has changed. The names of some fields have also changed. Nothing else has changed.